changes to board, node, and api
created function createNewBoard in the board class that generates a new
board. not finished since we have not finalised the map. also we need to
decide how we will generate the troops on the map.
i also made changes to the node class. created a function
generateAdjacentNodes which generates the adjacent nodes based on the
current node object. this also will need finishing since the map hasnt
been completed.
iv added a case for the new board command to the api. i added the new
board object to a session attribute and gave the response back to the
client in a JSON array.
This will all need to be discussed tomorrow as the board and node class
changes will need to be merged for you guys to work on if you agree with
the changes.
Sorry for long commit message :P