progress with gson
able to return new board json using the gson library. have got a jsp
page working when you can enter commands into a form and get a response.
currently only newboard is implemented. need to work on finding a way to
decode the json that is being stored in the session.