Changed attack, fortify reinforce-setup stage, added utility functions
Changed attack to now take 2 dice rolls instead of the army sizes also
added troops back into the territory if player can not take control.
Changed fortify so that only one territory can be fortified in a turn
and made it so the territories had to be adjacent.
Changed newTurn so while in setup-reinforce stage it checks if the
player has an army, if not changes to one who does.
Added utility functions getGameData and calcArmySize and changed
rollDice from private to protected.