fixing bugs + jsp page
fixed a bug in the game class in the nextPlayer method where there was a
call to indexOf. the equals method in player class needed to be
implemented. fixed the arrangement of paramters in the mainservlet claim
territory (my bad). working on getting the jsp page working with html
forms and jquery